pyproject.toml
[tool.poetry]
name = "roombooker"
version = "0.1.0"
description = ""
authors = ["Jamoliddin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
